- use size_t instead of unsigned long for return type of std::string::rfind
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@9677 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-core/src/PluginLibraryLoader.cpp
@@ -196,7 +196,7 @@ bool PluginLibraryLoader::initPluginDir(PluginLoader *loader) {
     while (success) {
       std::string currentPluginLibrary = pluginPath +"/"+ findData.cFileName;
       std::string lib(findData.cFileName);
-      unsigned long idx = lib.rfind('-');
+      size_t idx = lib.rfind('-');
 
       if (idx != std::string::npos) {
         std::string tulip_release(TULIP_MM_RELEASE);
